move interface to public utils and fix some build errors
