+++ packages/drivers/odsp-driver/src/odspPublicUtils.ts
@@ -4,8 +4,25 @@
  */
 
 import { hashFile, IsoBuffer } from "@fluidframework/common-utils";
+import { ISequencedDocumentMessage, ISnapshotTree } from "@fluidframework/protocol-definitions";
 
 export async function getHashedDocumentId(driveId: string, itemId: string): Promise<string> {
     const buffer = IsoBuffer.from(`${driveId}_${itemId}`);
     return encodeURIComponent(await hashFile(buffer, "SHA-256", "base64"));
+}
+
+export interface ISnapshotContents {
+    snapshotTree: ISnapshotTree;
+    blobs: Map<string, ArrayBuffer>;
+    ops: ISequencedDocumentMessage[];
+
+    /**
+     * Sequence number of the snapshot
+     */
+    sequenceNumber: number | undefined;
+
+    /**
+     * Sequence number for the latest op/snapshot for the file in ODSP
+     */
+    latestSequenceNumber: number | undefined;
 }
